requirements.txt
# Core LLM
ollama>=0.1.0
groq>=0.4.0

# Memory Framework
letta>=0.3.0

# Embedding + Vector DB
FlagEmbedding>=1.2.0
sentence-transformers>=2.2.0
chromadb>=0.4.0

# Knowledge Graph
neo4j>=5.0.0

# Agent Orchestration
langgraph>=0.1.0
langchain>=0.2.0

# Backend API
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0
websockets>=12.0

# Document collection
pypdf>=4.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
arxiv>=2.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
